Generate public/private keypairs for 3pid invites
Optionally include private keys in emails.

Perpetually serve up 'yes of course it's valid' to any public keys we
have issued, and hand them to the invite storer.

The fact that this currently re-implements truncated base64 encoding is
sad but we need it to be urlsafe and apparently we don't do that
anywhere else.