Accept JSON on all client-facing endpoints
In addition to x-form-www-urlencoded, based on what content type
the client sends. This will allow us to migrate to JSON to be
consistent with the client/server API.

Also remove support for camelcase parameters like sendAttempt and
clientSecret as all clients moved over to using the underscored
versions long ago so there should be no clients sending camelcase
in the wild by now.