Keep a year(ish) worth of logs
RotatingFileHandler doesn't have a 'keep everything' mode (because
it renames (.1 -> .2, .2 -> .3 etc, so it would be an arbitrarily
large number of reanmes). Onlyt keeping 3 days of logs isn't
sensible though, so keep a year's worth. The sensible thing to do
would be to employ a log archiver to archive the old ones.