Change lookup_hash index from medium to mxid (#222)
The lookup v2 endpoint was slow due to it not using an index. This is because the indexed column to lookup_hash was `medium`, but we select `mxid`.

Additionally, it was confusing that the `*.sql` files were not representative of the latest state (the migration files are). So, add a note in each file to reflect this and hopefully save future troubleshooter's time.