[DINSIC] Prevent a single failure aborting the federation loop (#225)
The DINSIC version of #141. The `dinsic` branch has the ability to replicate more than just associations to other sydent's. It can also send invite tokens and ephemeral keys. Thus, the code is structured in a bit of a different way where a dict containing all of these are sent (whereas in the `master` branch version, only associations are sent, which isn't wrapped in a bigger dict).

Also DINSIC's doesn't seem to work with matrix-is-tester, which is very sad indeed.