Add an option, always_obfuscate, to control whether parts of email addresses should always be redacted (#322)
When generating a third-party invite, Sydent tries to obfuscate part of the invited email address before creating and sending off an event that will eventually be inserted into the relevant room. This address is obfuscated so that it can both:

* Be recognised by whomever created it and additionally distinguished from other third-party invites by all users.
* Hide the full address in order to prevent leaking private information.

A question arises when a user has an extremely short email address, such as `abc@ex.org`. Sydent operators can configure the minimum amount of characters that should be shown, but what if that value is relatively high? The full address would be shown.

To combat this, Sydent has some code that will fire if the username or domain of the email address is shorter in length than that configured amount of characters. It's some simply logic, that says if the string is less than 5 characters, only reveal 3... if it's shorter than 3 characters, only reveal 1 etc.

https://github.com/matrix-org/sydent/blob/5a874b1951fb09fb8bacc05cc50ce509c2f4e362/sydent/http/servlets/store_invite_servlet.py#L170-L175

DINUM don't want this functionality however, and would rather have the whole address be shown. Thus, a config option is being added here to forgo this functionality if the operator would like.