Add an option, third_party_invite_username_separator_string, that allows redacting separate portions of a third party invite address (#323)
When generating a third-party invite, Sydent tries to obfuscate part of the invited email address before creating and sending off an event that will eventually be inserted into the relevant room. This address is obfuscated so that it can both:

* Be recognised by whomever created it and additionally distinguished from other third-party invites by all users.
* Hide the full address in order to prevent leaking private information.

The username and domain in an email address are obfuscated separately. However, this can pose an issue when emails look something like "John.Smith@example.com". This could be obfuscated to "John.Sm...@exam...".

While this has hidden the user's last name, their first name is still visible!

This PR adds an option that lets you configure a separator for email address usernames. The resulting strings after splitting the username using this separator will be obfuscated individually. So, if this option were set to `.`, then "John.Smith@example.com" would become, say, "Jo...Sm...@exam...".

Note that there currently only exists an option for usernames, as that's all that was requested.