Replace the triple-dot character with an ellipsis in third-party invites
Besides a slightly nicer presentation, this can be useful when email addresses
that contain period characters are obfuscated. Consider the address:
`firstname.lastname@example.com`. If this address ends up being obfuscated to
`firstname....@exa...`, we end up with a slightly odd-looking sequence of
four periods on the left-hand side.

With an ellipsis character, we'd instead get `firstname.…@exa…`, which is still
a little odd, but is clearer with regards to what exactly was obfuscated.