Change out the triple-dot sequence in third-party invites to an ellipsis character (#324)
Besides a slightly nicer presentation, this can be useful when email addresses that contain period characters are obfuscated. Consider the address: `firstname.lastname@example.com`.

If this address is obfuscated to `firstname....@exa...`, we end up with a slightly odd-looking sequence of four periods on the left-hand side.

With an ellipsis character, we'd instead get `firstname.…@exa…`, which is still a little odd, but is clearer with regards to what exactly was obfuscated.

Applying this to the `dinsic` branch as it grew out of [a dinsic PR conversation](https://github.com/matrix-org/sydent/pull/323#discussion_r526854002), but will be ported to mainline with the rest of the obfuscated code it touches.