Move crypto config handling over to SydentConfig
The SydentEd25519 class is no longer needed as all it did was
handle the signingkey config option before (which is now done
in CryptoConfig)

In the current setup, if no key is present a new one is generated
and saved back to file. This is a special case for the crypto config
and so the line `return self.crypto.save_key` isn't more general.

In the future this might not be needed (e.g. by having sydent only
do key generation when it first creates a config file for you, rather
than everytime one is not found) and so this special casing can be
removed.