Make the casefolding migration script more reliable (#379)
If the same MXID is associated with `Alice@example.com` and `alice@example.com`, we shouldn't send them an email.
While I was there, I also realised that the `mxid` variable for the template was set to the string `"mxid"` and the email subject was a bit obscure and not easy to change, so I've also fixed that.

I've also added exponential backoff when sending emails, so an error (e.g. a network blip, the SMTP server rate-limiting, etc) would not mean we need to restart the script from scratch.