Use `Response` instead of `IResponse`
Rationale: I want to annotate one of these, so that we know the type of
headers (which would allow us to detect the problem fixed by #415). if I
write a stub for `IResponse` I'll have to write a stub for the entire
`iweb` interface module, which seems like a PITA.

Also: AFAICS `Response` is the only implementation of `IResponse`. So I
feel a bit naughty, but not too naughty.