Catch failures to decode JSON in /register
Like #456, but for another error case.

If the remote homeserver fails to give us sensible response, that's not
a problem in our application. We can continue to serve everyone else. We
should tell the requester what happened and log a warning, but not an
error to Sentry.

Most of the examples I saw in Sentry were 404 responses serving up a
generic "page not found" body.