Handle failures to contact homeserver in Threepidunbind
Addresses
https://sentry.matrix.org/sentry/sydent/issues/235004/?query=is%3Aunresolved%20event.timestamp%3A%3E%3D2021-11-03T10%3A00%3A00

That issue is rare---it's only occurred twice---but since it's so close
to #456 I thought I may as well throw it in too.

(I suppose this might encounter the problem in #463 as well, though we
haven't seen it in Sentry. I think unbinds are just much less used that
registers. I don't know if there's a better way to handle all this
duplication though?)